+++ crates/cheatnet/tests/cheatcodes/cheat_execution_info.rs
@@ -139,6 +139,7 @@ struct TxInfo {
     pub nonce_data_availability_mode: Felt,
     pub fee_data_availability_mode: Felt,
     pub account_deployment_data: Vec<Felt>,
+    pub proof_facts: Vec<Felt>,
 }
 
 impl TxInfo {
@@ -172,6 +173,7 @@ impl TxInfo {
             nonce_data_availability_mode: clone_field!(nonce_data_availability_mode),
             fee_data_availability_mode: clone_field!(fee_data_availability_mode),
             account_deployment_data: clone_field!(account_deployment_data),
+            proof_facts: clone_field!(proof_facts),
         }
     }
 
@@ -253,6 +255,10 @@ fn start_cheat_execution_info_multiple_times() {
             contract_address,
             vec![Felt::from(777), Felt::from(888), Felt::from(999)],
         ),
+        proof_facts: operation_start(
+            contract_address,
+            vec![Felt::from(13), Felt::from(14), Felt::from(15)],
+        ),
     };
 
     let expected_tx_info = TxInfo::apply_mock_fields(&initial_tx_info_mock, &tx_info_before);
@@ -269,6 +275,7 @@ fn start_cheat_execution_info_multiple_times() {
         tip: Operation::Retain,
         nonce_data_availability_mode: Operation::Retain,
         account_deployment_data: Operation::Retain,
+        proof_facts: Operation::Retain,
         ..initial_tx_info_mock
     };