+++ crates/sncast/src/helpers/mod.rs
@@ -12,6 +12,7 @@ pub mod fee;
 pub mod interactive;
 pub mod ledger;
 pub mod output_format;
+pub mod proof;
 pub mod rpc;
 pub mod scarb_utils;
 pub mod signer;