+++ crates/sncast/src/starknet_commands/get/mod.rs
@@ -3,7 +3,9 @@ use sncast::helpers::configuration::CastConfig;
 use sncast::response::ui::UI;
 
 pub mod balance;
+pub mod class_hash_at;
 pub mod nonce;
+pub mod transaction;
 pub mod tx_status;
 
 #[derive(Args)]
@@ -18,20 +20,30 @@ pub enum GetCommands {
     /// Get the status of a transaction
     TxStatus(tx_status::TxStatus),
 
+    /// Get the transaction by hash
+    Tx(transaction::Transaction),
+
     /// Fetch balance of the account for specified token
     Balance(balance::Balance),
 
     /// Get nonce of a contract
     Nonce(nonce::Nonce),
+
+    /// Get class hash of a contract at a given address
+    ClassHashAt(class_hash_at::ClassHashAt),
 }
 
 pub async fn get(get: Get, config: CastConfig, ui: &UI) -> anyhow::Result<()> {
     match get.command {
         GetCommands::TxStatus(status) => tx_status::tx_status(status, config, ui).await?,
 
+        GetCommands::Tx(tx) => transaction::transaction(tx, config, ui).await?,
+
         GetCommands::Balance(balance) => balance::balance(balance, config, ui).await?,
 
         GetCommands::Nonce(nonce) => nonce::nonce(nonce, config, ui).await?,
+
+        GetCommands::ClassHashAt(args) => class_hash_at::class_hash_at(args, config, ui).await?,
     }
 
     Ok(())